review fix, fixture moved to conftest
